+++ modules/project-setting/init.js
