+++ modules/swaggerAPI/init.js
@@ -32,7 +32,7 @@ exports.init = (app, apiUrl) => {
             "./modules/company/routes.js",
             './modules/notification/prepare-notification-data/routes.js',
             "./modules/notification/notification-middleware/routes.js",
-            "./modules/projectSetting/routes.js",
+            "./modules/project-setting/routes.js",
             "./modules/trackerDownload/routes.js",
             "./modules/taskIndex/routes.js",
             "./modules/create-project/routes.js",
